+++ packages/terra-layout/CHANGELOG.md
@@ -3,6 +3,8 @@ Changelog
 
 Unreleased
 ----------
+### Changed
+* Added focus management to shift between layout panel and toggle button when it is opened and closed
 
 2.18.0 - (October 16, 2018)
 ------------------